requirements.txt
flask-mail
flask
flask-socketio
supabase
python-dotenv
requests
gunicorn
eventlet
flask-cors
flask-login
flask-sqlalchemy
sqlalchemy
psycopg2-binary
werkzeug
stripe
flask-WTF
flask-migrate
flask-bcrypt
flask-dance
pyngrok
flask-babel
flask_jwt_extended
openai
apscheduler
segno
